Add ldflags override for hub/tag in controller
