+++ cmd/mesh/profile-dump.go
@@ -15,10 +15,8 @@
 package mesh
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
-	"text/template"
 
 	"istio.io/pkg/version"
 
@@ -147,7 +145,7 @@ func genICPS(inFilename, profile, setOverlayYAML string, force bool, l *logger) 
 	tag := version.DockerInfo.Tag
 
 	if hub != "unknown" && tag != "unknown" {
-		buildHubTagOverlayYAML, err := generateHubTagOverlay(hub, tag)
+		buildHubTagOverlayYAML, err := helm.GenerateHubTagOverlay(hub, tag)
 		if err != nil {
 			return "", nil, err
 		}
@@ -284,34 +282,4 @@ func getConfigSubtree(manifest, path string) (string, error) {
 		return "", err
 	}
 	return string(out), nil
-}
-
-// generateHubTagOverlay creates an IstioControlPlaneSpec overlay YAML for hub and tag.
-func generateHubTagOverlay(hub, tag string) (string, error) {
-	hubTagYAMLTemplate := `
-hub: {{.Hub}}
-tag: {{.Tag}}
-`
-	ts := struct {
-		Hub string
-		Tag string
-	}{
-		Hub: hub,
-		Tag: tag,
-	}
-	return renderTemplate(hubTagYAMLTemplate, ts)
-}
-
-// helper method to render template
-func renderTemplate(tmpl string, ts interface{}) (string, error) {
-	t, err := template.New("").Parse(tmpl)
-	if err != nil {
-		return "", err
-	}
-	buf := new(bytes.Buffer)
-	err = t.Execute(buf, ts)
-	if err != nil {
-		return "", err
-	}
-	return buf.String(), nil
 }
