+++ pkg/helm/helm.go
@@ -15,7 +15,9 @@
 package helm
 
 import (
+	"bytes"
 	"fmt"
+	"html/template"
 	"io/ioutil"
 	"path/filepath"
 	"sort"
@@ -170,6 +172,36 @@ func OverlayYAML(base, overlay string) (string, error) {
 	}
 
 	return string(my), nil
+}
+
+// GenerateHubTagOverlay creates an IstioControlPlaneSpec overlay YAML for hub and tag.
+func GenerateHubTagOverlay(hub, tag string) (string, error) {
+	hubTagYAMLTemplate := `
+hub: {{.Hub}}
+tag: {{.Tag}}
+`
+	ts := struct {
+		Hub string
+		Tag string
+	}{
+		Hub: hub,
+		Tag: tag,
+	}
+	return renderTemplate(hubTagYAMLTemplate, ts)
+}
+
+// helper method to render template
+func renderTemplate(tmpl string, ts interface{}) (string, error) {
+	t, err := template.New("").Parse(tmpl)
+	if err != nil {
+		return "", err
+	}
+	buf := new(bytes.Buffer)
+	err = t.Execute(buf, ts)
+	if err != nil {
+		return "", err
+	}
+	return buf.String(), nil
 }
 
 // DefaultFilenameForProfile returns the profile name of the default profile for the given profile.
